+++ emodel/LinkDef.hh
@@ -72,8 +72,8 @@
 #pragma link C++ class std::pair<unsigned short,unsigned short>;
 
 //! links for hitmap
-#pragma link C++ class std::vector<std::map<int,double> >;
-#pragma link C++ class std::map<int,double>;
+#pragma link C++ class std::vector<std::map<int,float> >;
+#pragma link C++ class std::map<int,float>;
 
 //! links for pairs are requested by art (warnings) !!!!???!!??!!!!!--
 #pragma link C++ class std::pair<gate::BHit*,gate::BHit*>;
